+++ React_Template/src/util/ValidationUtil.js
@@ -1,4 +1,8 @@
 
+// 驗證為空值
+export const isEmpty = (val) => {
+    return !notEmpty(val);
+};
 
 // 驗證不為空值
 export const notEmpty = (val) => {