+++ React_Template/CHANGELOG.md
@@ -106,7 +106,7 @@ const localizer = momentLocalizer(moment);
 ### Bug fixing
 - Renamed `src/layouts/Dashboard/Dashboard.jsx` to `src/layouts/Admin.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Dashboard.jsx`
-- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/DrinkOrderList.jsx`
+- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/DrinksOrderList.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Typography.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/DrinksOrderInfo.jsx`
 - Renamed `src/components/Header` to `src/components/Navbars`